Use build tags to differentiate between features
