+++ configuration.json
@@ -4,13 +4,13 @@
 			"name" : "smart2", 
 			"width" : 200, 
 			"height" : 150,
-			"type" : "smartcrop"
+			"type" : "crop"
 		}, 
 			{
 			"name" : "smart0", 
 			"width" : 45, 
 			"height" : 35,
-			"type" : "smartcrop"
+			"type" : "crop"
 		}, 
 			{
 			"name" : "45x35", 
@@ -22,7 +22,7 @@
 			"name" : "smart1", 
 			"width" : 294, 
 			"height" : 300, 
-			"type" : "smartcrop"
+			"type" : "crop"
 		},
 	{
 			"name" : "294x300",
