+++ server/resizer/smartcrop.go
